- Refactored code for CustomDatePicker widget
